requirements.txt
asgiref==3.5.0
cloudinary==1.29.0
Django==3.2
django-cloudinary-storage==0.3.0
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
